поправлены недочеты
