+++ src/components/Message/Message.css
@@ -1,0 +1,7 @@
+.message {
+    align-self: flex-end;
+}
+
+.messageGood {
+    align-self: flex-start;
+}
